Refactor and add TwoSum functionality
Removed unnecessary using directives in PalindromeNumber.cs and SerchInsertPosition.cs.
Created a new TwoSumNumber class with a method to find indices of two numbers that add up to a target.
Updated Program.cs to include a call to the TwoSum method.
Note: Syntax errors in the TwoSum method need to be addressed.